Add Video Upload Part - Add ref at user & video schema - Update some views

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -21,13 +22,22 @@ export const getUpload = (req, res) =>
   res.render("upload", { pageTitle: "Upload Video" });
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
+      fileUrl,
+      owner: _id,
       title,
       description,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
   } catch (error) {
     console.log(error);
     return res.status(400).render("upload", { pageTitle: "Upload Video" });
@@ -38,7 +48,7 @@ export const postUpload = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video)
     return res.status(404).render("404", { pageTitle: "Video not found." });
 
@@ -50,6 +60,9 @@ export const getEdit = async (req, res) => {
   const video = await Video.findById(id);
   if (!video)
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  if (video.owner !== String(req.session.user._id)) {
+    return res.status(403).redirect("/");
+  }
 
   res.render("edit", { pageTitle: `Editing ${video.title}`, video });
 };
@@ -59,6 +72,9 @@ export const postEdit = async (req, res) => {
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) return res.render("404", { pageTitle: "Video not found." });
+  if (video.owner !== String(req.session.user._id)) {
+    return res.status(403).redirect("/");
+  }
 
   await Video.findByIdAndUpdate(id, {
     title,
@@ -71,6 +87,12 @@ export const postEdit = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video)
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  if (video.owner !== String(req.session.user._id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
 
   return res.redirect("/");
